balance(+-), settings panel, difficult handler (ploho)

+++ Assets/Scripts/Player/IDifficultHandler.cs
@@ -4,7 +4,7 @@ public interface IDifficultHandler
 {
     float ObstaclesQuantity { get; set; } 
     
-    float ObstacleHitPoints { get; set; }
+    int ObstacleHitPoints { get; set; }
     
     float SpawnRate { get; set; }
